requirements.txt
Flask==2.3.6
tensorflow==2.14.0
numpy==1.25.2
Pillow==10.1.0
matplotlib==3.8.0
scikit-learn==1.3.0
